Rename Contain with predicate to Any
